VPD-1392: Another small refactor
