+++ app/navigation/ReturnsNavigator.scala
@@ -21,6 +21,7 @@ import controllers.routes
 import models.*
 import models.returns.ReturnsUserAnswers
 import pages.*
+import pages.returns.{DeclareDutyPage, EnterDutyAmountPage}
 import play.api.Logging
 import play.api.http.HttpVerbs.GET
 import play.api.mvc.Call
